==> lab3-image-caption-generator/13-clear-translated-data.ipynb <==
import pandas as pd

path_captions = "D:/YandexDisk/datasets/russian-flickr-30k-images-metadata.csv"
df = pd.read_csv(path_captions, delimiter='|')
df.rename(columns={'image_name': 'image_name', ' comment_number': 'comment_number', ' comment': 'comment'}, inplace=True)
df.head(5)

# Проверка

– числа
– кавычки
– апострофы
– удалить точки

def is_there_number(sentence):
    for i in sentence.split():
        if i.isdigit():
            return True

    return False

def is_there_english(sentence):
    sentence = sentence.lower()

    alphabet = set('abcdefghijklmnopqrstuvwxyz')
    if any((char in alphabet) for char in sentence):
        return True
    else:
        return False

idxs_to_del = []

for idx in df.index:
    curr_str = df.iat[idx, 2]

    if is_there_english(curr_str):
        idxs_to_del.append(idx)
    elif is_there_number(curr_str):
        idxs_to_del.append(idx)
    elif curr_str.find('"') != -1:
        idxs_to_del.append(idx)
    elif curr_str.find("'") != -1:
        idxs_to_del.append(idx)

    if curr_str.find('.') != -1:
        df.iat[idx, 2] = curr_str.replace(".", "")

## Удаление

– числа
– кавычки
– апострофы (время)
– английский буквы (имена собственные)

df_to_delete = df.loc[idxs_to_del, :]
images_to_delete = df_to_delete["image_name"].unique()

print(len(images_to_delete))
print(df.shape)

df = df[~df.image_name.isin(images_to_delete)]
df.shape

## Сохранить

df.to_csv("D:/temp-russian-flickr-30k-images-metadata.csv", sep='|', encoding='utf-8', index=False)
